Remove multiple gsub calls
There were several places where `gsub` was being chained on a string
to do replacement of extra characters and the string format tokens.
I've pulled the multiple regexes up into several constants and chosen
to use the block syntax of `gsub` to provide find the replacements -
some of which are global in nature.

Renamed a constant or two, added some comments to help "future me" and
added a test for `::normalize`.